Project .c/.h files
Changed the structure so that the user will mainly work on producing a <project>.c/.h file when adding to the code. Idea is that user will define methods based off examples in default and the provided ATLAS.c and add simple "else if" statements in the source code to integrate their changes.